+++ game/rolls.py
@@ -1,20 +1,21 @@
-from django.core.exceptions import ObjectDoesNotExist
-
-from character.exceptions import AbilityNotFound
-from character.models.abilities import AbilityType
+from character.models.abilities import AbilityType, Ability
 from character.models.character import Character
 from utils.dice import Dice
 
 from .constants.events import RollResult
 from .models.events import RollRequest
+from .exceptions import RollInvalid
 
 
 def _roll(character: Character, ability_type: AbilityType) -> int:
+    """
+    Perform a roll and add proficiency bonus to the score,
+    if the character is proficient in the ability passed as argument.
+    """
     try:
         ability = character.abilities.get(ability_type=ability_type)
-    except ObjectDoesNotExist as exc:
-        raise AbilityNotFound(ability_type) from exc
-
+    except Ability.DoesNotExist:
+        raise RollInvalid(f"[{character}] does not have the ability: {ability_type}")
     score = Dice("d20").roll(ability.modifier)
     if character.is_proficient(ability):
         score += character.proficiency_bonus
@@ -25,7 +26,9 @@ def perform_roll(
     character: Character, request: RollRequest
 ) -> tuple[int, tuple[str, str]]:
     """
-    Perform dice roll.
+    Perform dice roll, according to DnD rules:
+    - It adds proficiency bonus in case the character has the ability given in the roll request.
+    - It adds or remove points in case the character has advantages or disadvantages.
 
     Args:
         character (Character): The character who performs the roll.
@@ -36,7 +39,6 @@ def perform_roll(
     """
 
     score = _roll(character, request.ability_type)
-
     has_advantage = character.has_advantage(request.roll_type, request.against)
     has_disadvantage = character.has_disadvantage(request.roll_type, request.against)
     if has_advantage and has_disadvantage:
@@ -48,7 +50,6 @@ def perform_roll(
             score = max(score, new_score)
         if has_disadvantage:
             score = min(score, new_score)
-
     if score >= request.difficulty_class:
         return score, RollResult.SUCCESS
     return score, RollResult.FAILURE