+++ CHANGELOG.md
@@ -5,10 +5,12 @@ Versions follow [Semantic Versioning](https://semver.org/) (`<major>.<minor>.<pa
 ## Unreleased
 
 ### Added
+* Add `delete_combats` command
 * DnD: Implement combat initiation
 * DnD: Implement combat structure
 
 ### Changed
+* Improve and rename some invoke commands
 * Enhance game events and their messages handling
 * Make celery tasks more robust
 * Refactor character app utils