+++ CHANGELOG.md
@@ -5,8 +5,8 @@ Versions follow [Semantic Versioning](https://semver.org/) (`<major>.<minor>.<pa
 ## Unreleased
 
 ### Added
+* DnD: Implement combat initialization
 * Add `delete_combats` command
-* DnD: Implement combat initiation
 * DnD: Implement combat structure
 
 ### Changed